+++ ExpressionTree/src/com/lukeherron/expressiontree/node/AlgebraicLeafNode.java
@@ -7,13 +7,13 @@ import com.lukeherron.expressiontree.visitor.Visitor;
  */
 public class AlgebraicLeafNode extends LeafNode {
 
-    private int item;
+    private double item;
 
     /**
      * Constructor for LeafNode. Accepts an int to be stored in this node
      * @param item The item to be added as the leaf node
      */
-    public AlgebraicLeafNode(int item) {
+    public AlgebraicLeafNode(double item) {
         this.item = item;
     }
 
@@ -22,14 +22,14 @@ public class AlgebraicLeafNode extends LeafNode {
      * @param item The item to be added as the leaf node
      */
     public AlgebraicLeafNode(String item) {
-        this.item = Integer.parseInt(item);
+        this.item = Double.parseDouble(item);
     }
 
     /**
      * Return the item stored in the node
      * @return int contained in this node
      */
-    public int item() {
+    public double item() {
         return this.item;
     }
