+++ ExpressionTree/src/com/lukeherron/expressiontree/node/CompositeAndNode.java
@@ -11,7 +11,7 @@ public class CompositeAndNode extends CompositeBinaryNode {
         super(left, right);
     }
 
-    public int item() {
+    public double item() {
         return '&';
     }
