+++ ExpressionTree/src/com/lukeherron/expressiontree/node/CompositeOrNode.java
@@ -11,7 +11,7 @@ public class CompositeOrNode extends CompositeBinaryNode {
         super(left, right);
     }
 
-    public int item() {
+    public double item() {
         return '|';
     }
